+++ lib/markdown-toolbar.rb
@@ -1,8 +1,6 @@
 require "markdown-toolbar/version"
 require "markdown-toolbar/engine"
 
-module Markdown
-  module Toolbar
-    require 'markdown-toolbar/engine'
-  end
+module MarkdownToolbar
+  require 'markdown-toolbar/engine'
 end